Finished version of the project
Because of all the error messages i was dealing with from getting blocked by the website, I switched to a different laptop to finish my project. After I completed it, I emailed the file to myself and saved it on my laptop so I could upload the file, but I accidentally uploaded the unfinished one that I had sent to the other laptop to complete. I was a bit out of it from staring at the computer for so long that I uploaded the file in the class module folder like I normally would, forgot that the finished version of it was in my downloads folder after saving it from my email. I just started up my laptop and realized my mistake, wanted to post the finished version of the program as well